🐛 fix(docs): fix manpage and add generation tool
The static RST manpage had broken section headers (all top-level `####`
overlines) producing empty `.TH` metadata and missing `.SH` sections.
It also only documented 6 of 11 commands and a subset of options.

Regenerate the RST with proper heading hierarchy from the argparse
parser via `tools/generate_manpage.py`. The build hook reads the static
file (avoiding runtime deps at build time), while tests verify the file
stays in sync with the actual CLI. Added changelog fragment.
